# swagger_output.json
{
  "swagger": "2.0",
  "info": {
    "version": "1.0.0",
    "title": "Scraper API",
    "description": "Documentation automatically generated by the <b>swagger-autogen</b> module."
  },
  "host": "localhost:3000",
  "basePath": "/",
  "tags": [
    {
      "name": "Sites",
      "description": "Endpoints"
    },
    {
      "name": "Models",
      "description": "Endpoints"
    }
  ],
  "schemes": [
    "http"
  ],
  "paths": {
    "/all": {
      "get": {
        "tags": [
          "Sites"
        ],
        "description": "",
        "parameters": [],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/search": {
      "get": {
        "tags": [
          "Sites"
        ],
        "description": "",
        "parameters": [
          {
            "name": "name",
            "in": "query",
            "type": "string"
          },
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "type": "object",
              "properties": {
                "site_name": {
                  "type": "string",
                  "example": "Somesite"
                }
              },
              "required": [
                "site_name"
              ]
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "201": {
            "description": "Return matched sites"
          }
        }
      }
    },
    "/create": {
      "post": {
        "tags": [
          "Sites"
        ],
        "description": "",
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "type": "object",
              "properties": {
                "site_name": {
                  "type": "string",
                  "example": "Somesite"
                },
                "url": {
                  "type": "string",
                  "example": "url"
                },
                "proxy": {
                  "type": "number",
                  "example": 0
                }
              },
              "required": [
                "site_name",
                "url",
                "proxy"
              ]
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Site created successfully",
            "schema": {
              "type": "object",
              "properties": {
                "url": {
                  "type": "string",
                  "example": "https://url.com"
                },
                "site_name": {
                  "type": "string",
                  "example": "site_name"
                },
                "locator": {
                  "type": "string",
                  "example": "/location.html"
                }
              },
              "xml": {
                "name": "main"
              }
            }
          }
        }
      }
    },
    "/delete": {
      "post": {
        "tags": [
          "Sites"
        ],
        "description": "",
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "type": "object",
              "properties": {
                "site_name": {
                  "type": "string",
                  "example": "site"
                }
              },
              "required": [
                "site_name"
              ]
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Site deleted successfully"
          }
        }
      }
    },
    "/all_models": {
      "get": {
        "tags": [
          "Models"
        ],
        "description": "",
        "parameters": [],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/search_model": {
      "post": {
        "tags": [
          "Models"
        ],
        "description": "",
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "type": "object",
              "properties": {
                "model": {
                  "type": "string",
                  "example": "Jhon Doe"
                },
                "site": {
                  "type": "number",
                  "example": 29
                }
              },
              "required": [
                "model",
                "site"
              ]
            }
          }
        ],
        "responses": {
          "200": {
            "description": "URL & locator returned successfully",
            "schema": {
              "type": "object",
              "properties": {
                "url": {
                  "type": "string",
                  "example": "https://url.com"
                },
                "site_name": {
                  "type": "string",
                  "example": "site_name"
                },
                "locator": {
                  "type": "string",
                  "example": "/location.html"
                }
              },
              "xml": {
                "name": "main"
              }
            }
          }
        }
      }
    },
    "/create_model": {
      "post": {
        "tags": [
          "Models"
        ],
        "description": "",
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "type": "object",
              "properties": {
                "name": {
                  "type": "string",
                  "example": "model_name"
                }
              },
              "required": [
                "name"
              ]
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Model created successfully"
          }
        }
      }
    },
    "/delete_model": {
      "post": {
        "tags": [
          "Models"
        ],
        "description": "",
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "type": "object",
              "properties": {
                "name": {
                  "type": "string",
                  "example": "modelA"
                }
              },
              "required": [
                "name"
              ]
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Site deleted successfully"
          }
        }
      }
    }
  }
}